test-requirements.txt
# Requirements for running the tests
pytest
py-cpuinfo
